chore: move *.vue to components/
