+++ frontend/src/index.ts
@@ -1,5 +1,5 @@
 import { createApp } from 'vue'
-import App from '@/App.vue'
+import App from '@/components/App.vue'
 
 const app = createApp(App)
 app.mount('#app')
